+++ src/main/java/com/example/info_coehgns/repository/ToDoRepository.java
@@ -3,5 +3,5 @@ package com.example.info_coehgns.repository;
 import com.example.info_coehgns.entity.ToDo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ToDoRepository extends JpaRepository<ToDo, Integer> {
+public interface ToDoRepository extends JpaRepository<ToDo, Long> {
 }
